+++ src/main/java/subway/application/service/line/LineQueryService.java
@@ -9,7 +9,6 @@ import subway.application.port.in.line.FindLineByIdUseCase;
 import subway.application.port.in.line.dto.response.LineQueryResponse;
 import subway.application.port.out.line.LoadLinePort;
 import subway.application.service.exception.NoSuchLineException;
-import subway.application.service.mapper.LineMapper;
 import subway.domain.line.Line;
 
 @Service
