+++ src/main/java/subway/application/service/route/RouteMapper.java
@@ -1,9 +1,10 @@
-package subway.application.service.mapper;
+package subway.application.service.route;
 
 import java.util.List;
 import java.util.stream.Collectors;
 import subway.application.port.in.route.dto.response.RouteQueryResponse;
 import subway.application.port.in.station.dto.response.StationQueryResponse;
+import subway.application.service.station.StationMapper;
 import subway.domain.fare.Fare;
 import subway.domain.route.Route;
