+++ src/main/java/subway/application/service/station/StationMapper.java
@@ -1,4 +1,4 @@
-package subway.application.service.mapper;
+package subway.application.service.station;
 
 import subway.adapter.out.persistence.entity.StationEntity;
 import subway.application.port.in.station.dto.response.StationQueryResponse;
